Add `og:site_name` meta tag
Customer complained about google using their domain as site name,
instead of a nice readable name. They use `og:site_name` meta tag as
a hint when generating site name. To increase the probability of having
a nice name we are adding this meta tag here.

Note that it intentionally uses the shop name instead of page_title as
it is expected to be consistent across all pages.